refactor: 边框12 tsx 改造
