+++ packages/datav-vue3/components/BorderBox12/index.ts
@@ -1,5 +1,5 @@
 import type { App, Plugin } from 'vue'
-import BorderBox12 from './src/index.vue'
+import BorderBox12 from './src/BorderBox12'
 
 export const BorderBox12Plugin: Plugin = {
   install(app: App) {
