+++ packages/docs/docs/Border/BorderBox12/demo.vue
@@ -1,9 +1,12 @@
 <template>
   <div demo-bg>
-    <dv-border-box-12>
+    <dv-border-box12>
       <div dv-bg>
         dv-border-box-12
       </div>
-    </dv-border-box-12>
+    </dv-border-box12>
   </div>
 </template>
+<script setup lang="ts">
+import { BorderBox12 as DvBorderBox12 } from '@kjgl77/datav-vue3'
+</script>
